Externalize main app
